+++ CHANGELOG.md
@@ -5,4 +5,10 @@
 ### First Release
 
 -   Release v1.0.0
--   Added functions - `formatDateTime` - Format a date.
+-   Added function - `formatDateTime` - Format a date.
+
+## 2.0.0
+
+### Major Release
+
+-   Modified function - `formatDateTime`. Removed `paddedValues` prop. For values without leading 0 new formats are added: h, H, m, s, S, SS, d, m